WIP: Loop table construction

+++ src/main/kotlin/tin/services/ontology/loopTable/LoopTable.kt
@@ -14,10 +14,10 @@ interface LoopTable<entryType: LoopTableEntry> {
      - efficient querying of entries. All 3 keys are required for each lookup
      */
 
-    val map: HashMap<entryType, Long>
+    val map: HashMap<entryType, Int>
 
-    fun get(entry: entryType): Long?
+    fun get(entry: entryType): Int?
 
-    fun set(entry: entryType, value: Long)
+    fun set(entry: entryType, value: Int)
 
 }
